refactor(capsule)!: target-gate the Wasmtime engine so the engine-agnostic machinery compiles for wasm32-unknown-unknown
The dispatcher, registry, access resolver, context, trait layer, and manifest
analysis (readiness/topic/toposort) in astrid-capsule are pure logic, but the
crate unconditionally compiled the Wasmtime engine and its native-only
dependencies (wasmtime/wasmtime-wasi/wasmparser, reqwest, socket2, tokio-util,
notify, nix, astrid-mcp, astrid-workspace, astrid-vfs, and tokio's net/
rt-multi-thread features), so no alternate engine host could reuse the shared
machinery for wasm32-unknown-unknown.

The Wasmtime engine, the process-backed MCP engine, manifest discovery, the
file-watch hot-reload path, and the per-Store StoreMemoryMeter are now gated to
native with cfg(not(all(target_arch = "wasm32", target_os = "unknown"))), and
their native-only dependencies move into a matching target-specific dependency
table. Portable modules route task spawning and timers through the astrid-runtime
facade. The kernel's cli_socket_listener and the context's overlay_registry keep
identical field arity on both targets via cfg'd type aliases (uninhabited on
wasm). astrid-capsule is added to the wasm portability CI guard. Native build/
test/clippy/fmt are byte-identical.

BREAKING CHANGE: the host audit-sink contract moved out of the engine subtree.
astrid_capsule::engine::wasm::host::audit_sink::{HostAuditSink, HostAuditEvent,
HostAuditOutcome} is now astrid_capsule::audit_sink::{...}. The crate-root
re-exports astrid_capsule::{HostAuditSink, HostAuditEvent, HostAuditOutcome}
(and CapsuleRuntimeLimits, HttpLimits) remain reachable and unchanged.

Closes #1142
