+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Changelog tracking starts with 0.2.0. Prior versions were not tracked.
 - **`astrid capsule check` — a static, CI-friendly linter for a capsule's tool wiring.** Several capsule-authoring mistakes fail *silently* at runtime: a `#[astrid::tool]` with no `[subscribe] "tool.v1.execute.<name>"` route advertises in `tools/list` but never executes (#1127); a tool capsule without the mandatory `[publish]` boilerplate can't return results or answer the describe fan-out; a `[subscribe]` handler that isn't `tool_execute_<name>` routes then gets denied. `astrid capsule check [PATH]` cross-checks the `#[astrid::tool("…")]` annotations in `src/` against the `Capsule.toml` `[subscribe]`/`[publish]` tables and reports each problem with the exact line to add, exiting non-zero on any finding so it drops straight into a CI job or pre-commit hook. It is deliberately **static** — it derives the advertised-tool set from source, not by instantiating the WASM component — so it needs no build, no daemon, no WASM runtime, and no `ASTRID_HOME`, making it fast and deterministic (`cargo check` for capsules). Check #1 reuses the same `tools_missing_execute_route` predicate as the kernel's load-time warning, so the CI gate and the runtime can never disagree; a `--deep` mode (ephemeral-load + the real `tool_describe`) is a documented later addition. Closes #1129.
 ### Changed
 
+- **`astrid-capsule`'s engine-agnostic machinery now compiles for `wasm32-unknown-unknown`, groundwork for an alternate capsule host.** The dispatcher, registry, access resolver, context, trait layer, and manifest analysis (readiness/topic/toposort) are pure logic, but the crate unconditionally compiled the Wasmtime engine and its native-only dependencies (`wasmtime`/`wasmtime-wasi`/`wasmparser`, `reqwest`, `socket2`, `tokio-util`, `notify`, `nix`, `astrid-mcp`, `astrid-workspace`, and `tokio`'s `net`/`rt-multi-thread` features), so no alternate engine host could reuse the shared machinery. The Wasmtime engine, the process-backed MCP engine, manifest discovery, the file-watch hot-reload path, and the per-Store `StoreMemoryMeter` are now target-gated to native (`cfg(not(all(target_arch = "wasm32", target_os = "unknown")))`), and their native-only dependencies move into a matching target-specific dependency table; the portable modules route task spawning and timers through the `astrid-runtime` facade. The host audit-sink contract (`HostAuditSink`, `HostAuditEvent`, `HostAuditOutcome`) — pure Rust with no Wasmtime dependency — moves out of the engine subtree to `astrid_capsule::audit_sink`; the crate-root re-exports (`astrid_capsule::{HostAuditSink, HostAuditEvent, HostAuditOutcome, CapsuleRuntimeLimits, HttpLimits}`) stay reachable and unchanged. An alternate host can now supply its own `ExecutionEngine` behind the same `dyn Capsule` trait. `astrid-capsule` is added to the `wasm32-unknown-unknown` portability CI guard; native `build`/`test`/`clippy`/`fmt` are unaffected. Closes #1142.
+
 - **The kernel's task-spawning and time surface is now selected per target through a new `astrid-runtime` facade crate, groundwork for alternate host profiles.** The kernel spawned tasks and read the clock via `tokio::spawn`, `tokio::time::{sleep, interval, timeout, Instant, MissedTickBehavior}`, `tokio::time::Instant`/`std::time::Instant`, and `std::time::SystemTime` directly at ~50 call sites — none of which exist on `wasm32-unknown-unknown` (no tokio runtime, no OS threads, no `std::time` clock), so a portable host could not reuse the kernel's scheduling and timekeeping. Those calls now route through `astrid-runtime`, a single seam that selects the surface at compile time: the native arm (`cfg(not(target_family = "wasm"))`) is pure `tokio`/`std` re-exports — every item resolves to the exact same type it replaced, so it is zero-cost with no behaviour change on native — while the wasm arm (`cfg(target_family = "wasm")`) maps the task surface to the JS microtask queue (`wasm-bindgen-futures`), the timer surface to JS timers (`wasmtimer`), and the wall clock to the browser time bridge (`web-time`). The facade is deliberately a `cfg`-gated re-export rather than a trait: the implementation never varies within one platform build, so there is no runtime polymorphism to model. `astrid-runtime` is added to the `wasm32-unknown-unknown` portability CI guard; native `build`/`test`/`clippy`/`fmt` are unaffected. See #1140.
 
 - **The kernel's pure-semantics crates now compile for `wasm32-unknown-unknown`, groundwork for a browser host profile.** `astrid-types`, `astrid-core`, `astrid-crypto`, `astrid-capabilities`, `astrid-audit`, `astrid-events`, `astrid-config`, and `astrid-approval` build for the wasm target with no behaviour change on native. Two structural blockers were cleared. First, the workspace-level `tokio` dependency pinned an additive base feature set (`net`, `signal`, `rt-multi-thread`) onto every `workspace = true` consumer — including leaf crates that only asked for `sync`/`time` — dragging in `mio` (which `compile_error!`s on wasm) and the multi-threaded runtime (which needs OS threads). The base is trimmed to the portable subset (`sync`, `macros`, `time`, `rt`, `io-util`); the native-only features are now declared locally by the crates that actually use them (uplink/cli/kernel/gateway/capsule for `net`, daemon/kernel for `signal`, daemon/emit/cli/capsule-install/capsule/hooks/kernel/gateway for `rt-multi-thread`). Second, `getrandom` (reached via `uuid`/`rand` at major 0.4 and `surrealkv` at major 0.3) refuses to build for wasm without a backend selection: the `wasm_js` feature is now enabled through `[target.'cfg(target_family = "wasm")'.dependencies]` entries on the crates owning those edges, paired with the build-time `--cfg getrandom_backend="wasm_js"` that `scripts/check-wasm-portability.sh` (a new `cargo check`-only regression guard, wired into CI) sets. The audit store's `block_in_place` fast path (native-only) is cfg-gated to its scoped-thread fallback on wasm. No public API changes; native `build`/`test`/`clippy`/`fmt` are unaffected.
