+++ crates/astrid-capsule/Cargo.toml
@@ -13,6 +13,10 @@ build = "build.rs"
 default = []
 fuzzing = []
 
+# Engine-agnostic (portable) dependencies. The dispatcher, registry, access
+# resolver, context, trait layer, and manifest analysis share these so an
+# alternate host (e.g. a browser WebAssembly build) can reuse the machinery
+# without the Wasmtime engine. Native-only deps live in the target table below.
 [dependencies]
 astrid-approval = { workspace = true }
 astrid-capabilities = { workspace = true }
@@ -20,47 +24,62 @@ astrid-capsule-types = { workspace = true }
 astrid-core = { workspace = true }
 astrid-crypto = { workspace = true }
 astrid-events = { workspace = true }
-astrid-mcp = { workspace = true }
+astrid-runtime = { workspace = true }
 astrid-storage = { workspace = true, features = ["keychain"] }
-astrid-vfs = { workspace = true }
-astrid-workspace = { workspace = true }
 anyhow = { workspace = true }
 arc-swap = { workspace = true }
 async-trait = { workspace = true }
 base64 = { workspace = true }
 blake3 = { workspace = true }
-bytes = "1"
 dashmap = { workspace = true }
 hex = { workspace = true }
 metrics = { workspace = true }
-notify = "8"
 parking_lot = { workspace = true }
 rand = { workspace = true }
+semver = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+sha2 = { workspace = true }
+tempfile = { workspace = true }
+thiserror = { workspace = true }
+# The portable manifest security gate parses URLs. `url` is wasm-clean; the
+# native HTTP host uses `reqwest::Url` (a re-export of this same type).
+url = { workspace = true }
+# Base tokio: the portable subset the workspace already provides (`sync`,
+# `macros`, `time`, `rt`, `io-util`). `net` + `rt-multi-thread` are native-only,
+# declared in the target table below.
+tokio = { workspace = true }
+toml = { workspace = true }
+tracing = { workspace = true }
+uuid = { workspace = true }
+
+# Native-only dependencies: the Wasmtime execution engine, its HTTP/socket/
+# process host functions, the MCP host engine, manifest discovery, and the
+# file-watch hot-reload path. Gated out of the `wasm32-unknown-unknown` browser
+# host profile, which supplies its own `ExecutionEngine`.
+[target.'cfg(not(all(target_arch = "wasm32", target_os = "unknown")))'.dependencies]
+astrid-mcp = { workspace = true }
+# `astrid-vfs` is native-only (it uses `cap-std` and tokio's filesystem
+# surface). Only `context.rs`'s `overlay_registry` field references it, behind
+# the `OverlayRegistry` alias that becomes an uninhabited type on wasm.
+astrid-vfs = { workspace = true }
+astrid-workspace = { workspace = true }
+bytes = "1"
+nix = { workspace = true }
+notify = "8"
 # `astrid:http@1.1.0` `auto-decompress` toggles gzip/brotli/deflate/zstd per
 # request; those builder methods are feature-gated in reqwest. Enabled here so
 # the host can honour (and disable) decompression. The workspace base dep stays
 # minimal; Cargo unions these features in for the capsule crate's build.
 reqwest = { workspace = true, features = ["gzip", "brotli", "deflate", "zstd"] }
-semver = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-sha2 = { workspace = true }
 socket2 = "0.6"
-tempfile = { workspace = true }
-thiserror = { workspace = true }
 # `net` (host net fns + socket handshake) and `rt-multi-thread`
 # (`block_in_place` host bridges): native-only, no longer in the tokio base.
 tokio = { workspace = true, features = ["net", "rt-multi-thread"] }
 tokio-util = { workspace = true }
-toml = { workspace = true }
-tracing = { workspace = true }
-uuid = { workspace = true }
 wasmparser = { workspace = true }
 wasmtime = { workspace = true }
 wasmtime-wasi = { workspace = true }
-
-[target.'cfg(unix)'.dependencies]
-nix = { workspace = true }
 
 [dev-dependencies]
 astrid-crypto = { workspace = true }
