+++ crates/astrid-capsule/src/audit_sink.rs
