+++ crates/astrid-capsule/src/dispatcher.rs
@@ -425,7 +425,7 @@ fn dispatch_to_capsule_queues(
     let topic_clone = Arc::clone(&topic);
     let ipc_clone = ipc_message.clone();
     let chain_locks_clone = Arc::clone(chain_locks);
-    tokio::task::spawn(async move {
+    astrid_runtime::spawn(async move {
         let mut current_payload = (*payload_bytes).clone();
 
         for (capsule, action) in &matches_owned {
@@ -586,7 +586,7 @@ fn get_or_spawn_consumer(
     let capsule_arc = Arc::clone(capsule);
     let queues_arc = Arc::clone(queues);
     let cleanup_key = effective_key.clone();
-    tokio::task::spawn(async move {
+    astrid_runtime::spawn(async move {
         run_consumer(rx, capsule_arc, queues_arc, cleanup_key).await;
     });
     tx
@@ -604,7 +604,7 @@ async fn run_consumer(
     key: (CapsuleId, PrincipalKey),
 ) {
     loop {
-        match tokio::time::timeout(idle_consumer_grace(), rx.recv()).await {
+        match astrid_runtime::time::timeout(idle_consumer_grace(), rx.recv()).await {
             Ok(Some(work)) => {
                 debug!(
                     capsule_id = %capsule.id(),
