+++ crates/astrid-capsule/src/engine/wasm/host/mod.rs
@@ -9,9 +9,8 @@
 
 /// Capsule-level approval requests.
 pub(crate) mod approval;
-/// Synchronous host-audit sink: routes sensitive per-action host calls
-/// (fs/net/process) to the kernel's signed, hash-chained audit log.
-pub mod audit_sink;
+/// Producer-side tests for the per-action host-audit seam (the sink trait
+/// itself lives at [`crate::audit_sink`]).
 #[cfg(test)]
 mod audit_sink_tests;
 /// Runtime operator-consent for local-egress (transport-origin gated).
