+++ scripts/check-wasm-portability.sh
@@ -28,6 +28,7 @@ CRATES=(
   astrid-approval
   astrid-capsule-types
   astrid-runtime
+  astrid-capsule
 )
 
 if ! rustup target list --installed 2>/dev/null | grep -qx "$TARGET"; then
